Fix a runtime bug after 20 years
I hardcoded a path to a font, into a location that is absolute and not
always available on users filesystem.

Actually - I do not depend on users having BorlandPascal7 installed.

The code tried to load a font from disk, instead I load the font from
BIOS and thats it. That font was the BIOS font anyway.. so its a
win-win scenario.